+++ e2e/react/vite.config.ts
@@ -1,8 +1,9 @@
 import react from '@vitejs/plugin-react'
+import adapter from 'houdini-adapter-cloudflare-workers'
 import houdini from 'houdini/vite'
 import { defineConfig } from 'vite'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-	plugins: [houdini(), react({ fastRefresh: false })],
+	plugins: [houdini({ adapter }), react({ fastRefresh: false })],
 })
