+++ packages/houdini-adapter-cloudflare-workers/src/index.ts
@@ -1,0 +1,7 @@
+import type { Adapter } from 'houdini'
+
+const adapter: Adapter = async ({ config, conventions }) => {
+	console.log('hello from adapter!')
+}
+
+export default adapter
