+++ packages/houdini-react/src/plugin/codegen/render.test.ts
@@ -1,8 +1,7 @@
-import { fs, parseJS } from 'houdini'
+import { fs, parseJS, routerConventions } from 'houdini'
 import { test, expect } from 'vitest'
 
 import { test_config } from '../config'
-import { render_server_path, render_client_path } from '../conventions'
 import { generate_renders } from './render'
 
 test('generates render functions', async function () {
@@ -10,14 +9,20 @@ test('generates render functions', async function () {
 
 	await generate_renders(config)
 
-	const render_client = await parseJS((await fs.readFile(render_client_path(config))) ?? '', {
-		plugins: ['jsx'],
-	})
+	const render_client = await parseJS(
+		(await fs.readFile(routerConventions.render_client_path(config))) ?? '',
+		{
+			plugins: ['jsx'],
+		}
+	)
 	expect(render_client).toMatchInlineSnapshot('')
 
-	const render_server = await parseJS((await fs.readFile(render_server_path(config))) ?? '', {
-		plugins: ['jsx'],
-	})
+	const render_server = await parseJS(
+		(await fs.readFile(routerConventions.render_server_path(config))) ?? '',
+		{
+			plugins: ['jsx'],
+		}
+	)
 	expect(render_server).toMatchInlineSnapshot(`
 		import React from "react";
 		import { renderToStream } from "react-streaming/server";
