+++ packages/houdini-react/src/plugin/codegen/typeRoot.ts
@@ -1,7 +1,6 @@
-import { type Config, path, fs } from 'houdini'
+import { type Config, path, fs, PageManifest, ProjectManifest } from 'houdini'
 
 import { dedent } from '../dedent'
-import type { PageManifest, ProjectManifest } from './manifest'
 
 export async function generate_type_root({
 	config,
