+++ packages/houdini/src/lib/router/conventions.ts
@@ -1,4 +1,4 @@
-import { path, fs, type Config } from 'houdini'
+import { path, fs, type Config } from '../'
 
 /** The location of the project's router */
 export function router_path(config: Config) {
@@ -15,7 +15,7 @@ export function render_client_path(config: Config, base?: string) {
 }
 
 export function render_server_path(config: Config, base?: string) {
-	return path.join(units_dir(config, base), 'render', 'server.jsx')
+	return path.join(units_dir(config, base), 'render', 'server.js')
 }
 
 export function render_app_path(config: Config, base?: string) {
