+++ packages/houdini/src/lib/router/index.ts
@@ -1,0 +1,11 @@
+import * as _conventions from './conventions'
+
+/**
+ * This directory contains all of the utilities that are designed to
+ * be used by plugins that generate router code.
+ */
+export * from './manifest'
+
+export const routerConventions = _conventions
+
+export * from './types'
