+++ packages/houdini/src/lib/router/types.ts
@@ -1,0 +1,8 @@
+import { routerConventions } from '.'
+import { Config } from '../config'
+
+export type Adapter = (args: {
+	config: Config
+	conventions: typeof routerConventions
+	sourceDir: string
+}) => void | Promise<void>
