+++ packages/houdini/src/vite/houdini.ts
@@ -1,11 +1,12 @@
 import type { SourceMapInput } from 'rollup'
-import type { Plugin as VitePlugin, UserConfig } from 'vite'
+import type { Plugin as VitePlugin, UserConfig, ResolvedConfig } from 'vite'
 
 import generate from '../codegen'
 import type { Config, PluginConfig } from '../lib'
-import { path, getConfig, formatErrors, deepMerge } from '../lib'
+import { path, getConfig, formatErrors, deepMerge, routerConventions } from '../lib'
 
 let config: Config
+let viteConfig: ResolvedConfig
 
 export default function Plugin(opts: PluginConfig = {}): VitePlugin {
 	return {
@@ -16,15 +17,15 @@ export default function Plugin(opts: PluginConfig = {}): VitePlugin {
 		enforce: 'pre',
 
 		// add watch-and-run to their vite config
-		async config(viteConfig, ...rest) {
+		async config(userConfig, ...rest) {
 			config = await getConfig(opts)
 
 			let result: UserConfig = {
 				server: {
-					...viteConfig.server,
+					...userConfig.server,
 					fs: {
-						...viteConfig.server?.fs,
-						allow: ['.'].concat(viteConfig.server?.fs?.allow || []),
+						...userConfig.server?.fs,
+						allow: ['.'].concat(userConfig.server?.fs?.allow || []),
 					},
 				},
 			}
@@ -42,6 +43,55 @@ export default function Plugin(opts: PluginConfig = {}): VitePlugin {
 			}
 
 			return result
+		},
+
+		async buildEnd(args) {
+			for (const plugin of config.plugins) {
+				if (typeof plugin.vite?.buildEnd !== 'function') {
+					continue
+				}
+
+				await plugin.vite!.buildEnd.call(this, args, config)
+			}
+		},
+
+		async configResolved(conf) {
+			viteConfig = conf
+			for (const plugin of config.plugins) {
+				if (typeof plugin.vite?.configResolved !== 'function') {
+					continue
+				}
+
+				await plugin.vite!.configResolved.call(this, conf)
+			}
+		},
+
+		// called when all of the bundles have been generated (ie, when vite is done)
+		// we use this to generate the final assets needed for a production build of the server.
+		// this is only called when bundling (ie, not in dev mode)
+		async closeBundle() {
+			for (const plugin of config.plugins) {
+				if (typeof plugin.vite?.closeBundle !== 'function') {
+					continue
+				}
+
+				await plugin.vite!.closeBundle.call(this)
+			}
+
+			// if we dont' have an adapter, we don't need to do anything
+			if (!opts.adapter) {
+				return
+			}
+
+			// tell the user what we're doing
+			console.log('🎩 Generating Server Assets...')
+
+			// invoke the adapter
+			await opts.adapter({
+				config,
+				conventions: routerConventions,
+				sourceDir: viteConfig.build.outDir,
+			})
 		},
 
 		// when the build starts, we need to make sure to generate
